sample/multi_agent_example/pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-texttospeech>=2.29.0",
    "google-generativeai>=0.8.5",
    "gtts>=2.5.4",
    "langchain>=0.3",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.7",
    "moviepy==1.0.3",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "streamlit>=1.49.1",
    "typer>=0.17.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.18.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.0",
]
